requirements2.txt
imageio==2.3.0
imageio-ffmpeg
matplotlib==2.2.2
numpy==1.15.0
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
torchvision==0.2.1
tqdm==4.24.0
moviepy
requests
Flask
flask_uploads
Werkzeug==0.16.0
flask_wtf
GitPython
#view issue: 'Fail to install torch==1.0.0 win10' https://github.com/AliaksandrSiarohin/first-order-model/issues/129
#pip install -f https://download.pytorch.org/whl/torch_stable.html torch==1.4.0